+++ src/app/recommend/page.tsx
@@ -3,11 +3,17 @@ import Button from "@/components/atoms/Button";
 import Icon from "@/components/atoms/Icon";
 import Text from "@/components/atoms/Text";
 import ButtonLabel from "@/components/molecules/ButtonLabel";
-import ButtonLabel_ing from "@/components/molecules/ButtonLabel_ing";
+import ButtonLabel_ing from "@/components/molecules/ButtonLabel";
 import Icon_Logo from "@/svg/Icon-logo";
 
 function RecommendPage() {
-  return <ButtonLabel_ing type="openStudy" />;
+  return (
+    <div className="w-[117px] h-[49px]">
+      <Button type="radius-none" theme="ordinary">
+        경기도
+      </Button>
+    </div>
+  );
 }
 
 export default RecommendPage;
